+++ CHANGELOG.md
@@ -50,6 +50,7 @@ CHANGELOG
 - Don't use Cerberus 1.3 because of https://github.com/pyeve/cerberus/issues/489
 
 ### Tools
+- intelmqdump: Fix creation of pipeline object by providing a logger.
 
 ### Contrib